Language_Translater_Using_LangChain_with_Gemini_LLM/requirements.txt
langchain-google-genai
pillow
python-dotenv
streamlit

